Fundamental in react components (#1793)
* Change fundamental versions

* Fix imports

* Fix catalog

* Fix catalog details crashing

* Fix FormInput in serviceInstanceList tests

* Fix card style

* Fix Card test

* Fix ServiceClassList test

* Update snapshots

* WIP

* Use react-shared spinner

* Fix ServiceInstanceTable test

* Uncomment test

* Fix ServiceInstanceHeader layout

* Use fd-react instead of react-components

* Fix test warnings

* Fix FilterDropdown display

* Moving tabs into react-shared in progress

* Use and simplify new Tabs

* Suppress Luigi warning in ModalWithForm test

* Add tab tests

* Fix styles

* Cleanup Tabs

* Cleanup Tabs usages

* Use react-shared modals

* Use shared tooltips

* Use GenericList instead of Table

* Fix catalog tests

* Fix instances list additional header column and bindings list row height

* Add tests for simpleDelete

* Unify className prop name

* Review fixes 1

* Fix instance parameter schema form

* Fix tabs callback

* Improve simpleDelete test

* Fix header after merge

* Refetch bindings manually

* Render both tabs and hide/show with css, fix tests

+++ shared/components/GenericList/actionHandlers/simpleDelete.js
@@ -9,34 +9,30 @@ function displayConfirmationMessage(entityType, entityName) {
         buttonConfirm: 'Delete',
         buttonDismiss: 'Cancel',
       })
-      .then(() => {
-        resolve();
-      })
-      .catch(e => {});
+      .then(() => resolve(true))
+      .catch(_e => resolve(false));
   });
 }
 
-export function handleDelete(
+export async function handleDelete(
   entityType,
   entityId,
   entityName,
   deleteRequestFn,
   callback = () => {},
 ) {
-  displayConfirmationMessage(entityType, entityName)
-    .then(() => {
-      return deleteRequestFn(entityId, entityName);
-    })
-    .then(() => {
+  try {
+    if (await displayConfirmationMessage(entityType, entityName)) {
+      deleteRequestFn(entityId, entityName);
       callback();
-    })
-    .catch(err => {
-      LuigiClient.uxManager().showAlert({
-        text: `An error occurred while deleting ${entityType} ${entityName}: ${err.message}`,
-        type: 'error',
-        closeAfter: 10000,
-      });
+    }
+  } catch (err) {
+    LuigiClient.uxManager().showAlert({
+      text: `An error occurred while deleting ${entityType} ${entityName}: ${err.message}`,
+      type: 'error',
+      closeAfter: 10000,
     });
+  }
 }
 
 export function easyHandleDelete(
@@ -48,22 +44,24 @@ export function easyHandleDelete(
   notificationManager,
   callback = () => {},
 ) {
-  displayConfirmationMessage(entityType, entityName)
-    .then(async () => {
-      try {
-        const result = await deleteRequestFn(deleteRequestParam);
-        const isSuccess =
-          result.data &&
-          (deleteRequestName ? result.data[deleteRequestName] : true);
-        if (isSuccess) {
-          notificationManager.notifySuccess({
-            content: `${entityName} deleted`,
-          });
-        } else {
-          throw Error();
+  return displayConfirmationMessage(entityType, entityName)
+    .then(async shouldDelete => {
+      if (shouldDelete) {
+        try {
+          const result = await deleteRequestFn(deleteRequestParam);
+          const isSuccess =
+            result.data &&
+            (deleteRequestName ? result.data[deleteRequestName] : true);
+          if (isSuccess) {
+            notificationManager.notifySuccess({
+              content: `${entityName} deleted`,
+            });
+          } else {
+            throw Error();
+          }
+        } catch (e) {
+          throw e;
         }
-      } catch (e) {
-        throw e;
       }
     })
     .then(() => {
